Amended debian/rules clean, now with less violence
The SConstruct file doesn't clean up a bunch of things, such as the
__pycache__ directories. This is fine, but they still gotta go away for
`dpkg-source` to be happy with the build results. Likewise, the scons
build state file needs to go away.

Previously, I was using `git clean -fxd` because it seemed unlikely that
anyone would be working on the package while also trying to rebuild
it... And then I did that. Less than an hour after that decision, and I
already vaporized some work. Big brain. (yay for vim .swp files, saved
my stuff).